+++ project.flutter/client_app/linux/flutter/generated_plugins.cmake
@@ -6,6 +6,7 @@ list(APPEND FLUTTER_PLUGIN_LIST
   gtk
   sw_rend
   url_launcher_linux
+  zstandard_linux
 )
 
 list(APPEND FLUTTER_FFI_PLUGIN_LIST