+++ project.flutter/client_app/windows/flutter/generated_plugin_registrant.cc
@@ -10,6 +10,7 @@
 #include <share_plus/share_plus_windows_plugin_c_api.h>
 #include <sw_rend/sw_rend_plugin_c_api.h>
 #include <url_launcher_windows/url_launcher_windows.h>
+#include <zstandard_windows/zstandard_windows_plugin_c_api.h>
 
 void RegisterPlugins(flutter::PluginRegistry* registry) {
   AppLinksPluginCApiRegisterWithRegistrar(
@@ -20,4 +21,6 @@ void RegisterPlugins(flutter::PluginRegistry* registry) {
       registry->GetRegistrarForPlugin("SwRendPluginCApi"));
   UrlLauncherWindowsRegisterWithRegistrar(
       registry->GetRegistrarForPlugin("UrlLauncherWindows"));
+  ZstandardWindowsPluginCApiRegisterWithRegistrar(
+      registry->GetRegistrarForPlugin("ZstandardWindowsPluginCApi"));
 }