+++ project.flutter/client_app/windows/flutter/generated_plugins.cmake
@@ -7,6 +7,7 @@ list(APPEND FLUTTER_PLUGIN_LIST
   share_plus
   sw_rend
   url_launcher_windows
+  zstandard_windows
 )
 
 list(APPEND FLUTTER_FFI_PLUGIN_LIST