+++ tool.dev/format_code.bash
@@ -4,6 +4,7 @@
 # brew install fd
 # brew install swiftformat
 # brew install uncrustify
+# dart ships with the Flutter SDK
 
 ROOT="$(readlink -f "$(dirname "${BASH_SOURCE[0]}")/..")"
 
@@ -12,3 +13,5 @@ cd "$ROOT"
 fd '\.swift$' project.ios/LowResRMX project.ios/Widgets | grep -v zstd | xargs -n1 swiftformat --quiet --swift-version 5.9 --allman true  --indent tab
 
 fd '\.[h|m|c]$' project.ios/LowResRMX project.ios/Widgets frontend.sdl backend.core | xargs uncrustify -q -c tool.dev/uncrustify.cfg --no-backup --replace -l OC
+
+dart format project.flutter/client_app/lib project.flutter/client_app/test project.flutter/core_plugin/lib project.flutter/core_plugin/android